refactor: move state actions from widget to index
- moves logic from SwapWidget to Swap index in order to share logic between both the price chart & swap widget.
